# perf: remove `<img>`'s `src` when offscreen