--- a/bin/bundlesize.js
+++ b/bin/bundlesize.js
@@ -5,7 +5,7 @@ import { promisify } from 'node:util'
 import prettyBytes from 'pretty-bytes'
 import fs from 'node:fs/promises'
 
-const MAX_SIZE_MIN = '37 kB'
+const MAX_SIZE_MIN = '38 kB'
 const MAX_SIZE_MINGZ = '13 kB'
 
 const FILENAME = './bundle.js'