This is an additional optimization on top of #445 in order to address #444.

It turns out that browsers seem to do some extra work when a `src` is set on an `<img>`, even if that image is offscreen, even if it's marked `<img loading=lazy>`, and _even_ if it's hidden with `content-visibility: auto`. To fix that, you can simply use the old `data-src` trick and set the `src` only when the `<img>` is visible.

To detect visibility, we can use the `contentvisibilityautostatechange` event (kind of like an `IntersectionObserver` but a bit simpler).

Oddly this event only seems to fire on the element itself and not bubble in Firefox only, but we can just set the event listener on the element itself to fix that.

Improvement in Chrome: 13-19%

```
┌─────────────┬─────────────────┐
│     Browser │ chrome-headless │
│             │ 128.0.0.0       │
├─────────────┼─────────────────┤
│ Sample size │ 300             │
└─────────────┴─────────────────┘

┌─────────────┬─────────────┬─────────────┬─────────────────────┬───────────────────┬───────────────────┐
│ Benchmark   │ Version     │ Bytes       │            Avg time │    vs this-change │    vs tip-of-tree │
│             │             │             │                     │                   │       tip-of-tree │
├─────────────┼─────────────┼─────────────┼─────────────────────┼───────────────────┼───────────────────┤
│ this-change │ <none>      │ 1029.47 KiB │ 155.77ms - 165.62ms │                   │            faster │
│             │             │             │                     │          -        │         13% - 19% │
│             │             │             │                     │                   │ 24.88ms - 37.22ms │
├─────────────┼─────────────┼─────────────┼─────────────────────┼───────────────────┼───────────────────┤
│ tip-of-tree │ tip-of-tree │ 1027.26 KiB │ 188.03ms - 195.47ms │            slower │                   │
│             │             │             │                     │         15% - 24% │          -        │
│             │             │             │                     │ 24.88ms - 37.22ms │                   │
└─────────────┴─────────────┴─────────────┴─────────────────────┴───────────────────┴───────────────────┘

```

Improvement in Firefox: 35-37%

```
┌─────────────┬──────────────────┐
│     Browser │ firefox-headless │
│             │ 130.0            │
├─────────────┼──────────────────┤
│ Sample size │ 200              │
└─────────────┴──────────────────┘

┌─────────────┬─────────────┬─────────────┬─────────────────────┬─────────────────────┬─────────────────────┐
│ Benchmark   │ Version     │ Bytes       │            Avg time │      vs this-change │      vs tip-of-tree │
│             │             │             │                     │                     │         tip-of-tree │
├─────────────┼─────────────┼─────────────┼─────────────────────┼─────────────────────┼─────────────────────┤
│ this-change │ <none>      │ 1020.95 KiB │ 240.02ms - 245.81ms │                     │              faster │
│             │             │             │                     │            -        │           35% - 37% │
│             │             │             │                     │                     │ 131.03ms - 139.76ms │
├─────────────┼─────────────┼─────────────┼─────────────────────┼─────────────────────┼─────────────────────┤
│ tip-of-tree │ tip-of-tree │ 1071.07 KiB │ 375.04ms - 381.58ms │              slower │                     │
│             │             │             │                     │           53% - 58% │            -        │
│             │             │             │                     │ 131.03ms - 139.76ms │                     │
└─────────────┴─────────────┴─────────────┴─────────────────────┴─────────────────────┴─────────────────────┘

```